Combine token card Activities and Fragments with non-token classes
