+++ judo-sdk/src/main/java/com/judopay/PaymentPresenter.java
@@ -4,6 +4,7 @@ import com.google.gson.Gson;
 import com.judopay.arch.Scheduler;
 import com.judopay.model.Card;
 import com.judopay.model.PaymentRequest;
+import com.judopay.model.TokenRequest;
 
 class PaymentPresenter extends BasePresenter {
 
@@ -35,6 +36,29 @@ class PaymentPresenter extends BasePresenter {
         }
 
         apiService.payment(builder.build())
+                .subscribeOn(scheduler.backgroundThread())
+                .observeOn(scheduler.mainThread())
+                .subscribe(callback(), error());
+    }
+
+    public void performTokenPayment(Card card, JudoOptions options) {
+        this.loading = true;
+        transactionCallbacks.showLoading();
+
+        TokenRequest tokenRequest = new TokenRequest.Builder()
+                .setAmount(options.getAmount())
+                .setCardAddress(card.getCardAddress())
+                .setCurrency(options.getCurrency())
+                .setJudoId(options.getJudoId())
+                .setYourConsumerReference(options.getConsumerRef())
+                .setCv2(card.getSecurityCode())
+                .setToken(options.getCardToken())
+                .setMetaData(options.getMetaDataMap())
+                .setEmailAddress(options.getEmailAddress())
+                .setMobileNumber(options.getMobileNumber())
+                .build();
+
+        apiService.tokenPayment(tokenRequest)
                 .subscribeOn(scheduler.backgroundThread())
                 .observeOn(scheduler.mainThread())
                 .subscribe(callback(), error());
