+++ judo-sdk/src/main/java/com/judopay/PreAuthPresenter.java
@@ -4,6 +4,7 @@ import com.google.gson.Gson;
 import com.judopay.arch.Scheduler;
 import com.judopay.model.Card;
 import com.judopay.model.PaymentRequest;
+import com.judopay.model.TokenRequest;
 
 class PreAuthPresenter extends BasePresenter {
 
@@ -35,6 +36,30 @@ class PreAuthPresenter extends BasePresenter {
         }
 
         apiService.preAuth(builder.build())
+                .subscribeOn(scheduler.backgroundThread())
+                .observeOn(scheduler.mainThread())
+                .subscribe(callback(), error());
+    }
+
+
+    public void performTokenPreAuth(Card card, JudoOptions options) {
+        this.loading = true;
+        transactionCallbacks.showLoading();
+
+        TokenRequest tokenRequest = new TokenRequest.Builder()
+                .setAmount(options.getAmount())
+                .setCardAddress(card.getCardAddress())
+                .setCurrency(options.getCurrency())
+                .setJudoId(options.getJudoId())
+                .setYourConsumerReference(options.getConsumerRef())
+                .setCv2(card.getSecurityCode())
+                .setEmailAddress(options.getEmailAddress())
+                .setMobileNumber(options.getMobileNumber())
+                .setMetaData(options.getMetaDataMap())
+                .setToken(options.getCardToken())
+                .build();
+
+        apiService.tokenPreAuth(tokenRequest)
                 .subscribeOn(scheduler.backgroundThread())
                 .observeOn(scheduler.mainThread())
                 .subscribe(callback(), error());
