--- judo-sdk/src/main/java/com/judopay/TokenPreAuthPresenter.java
@@ -1,37 +1,0 @@
-package com.judopay;
-
-import com.google.gson.Gson;
-import com.judopay.arch.Scheduler;
-import com.judopay.model.Card;
-import com.judopay.model.TokenRequest;
-
-class TokenPreAuthPresenter extends BasePresenter {
-
-    public TokenPreAuthPresenter(TransactionCallbacks callbacks, JudoApiService judoApiService, Scheduler scheduler, Gson gson) {
-        super(callbacks, judoApiService, scheduler, gson);
-    }
-
-    public void performTokenPreAuth(Card card, JudoOptions options) {
-        this.loading = true;
-        transactionCallbacks.showLoading();
-
-        TokenRequest tokenTransaction = new TokenRequest.Builder()
-                .setAmount(options.getAmount())
-                .setCardAddress(card.getCardAddress())
-                .setCurrency(options.getCurrency())
-                .setJudoId(options.getJudoId())
-                .setYourConsumerReference(options.getConsumerRef())
-                .setCv2(card.getSecurityCode())
-                .setEmailAddress(options.getEmailAddress())
-                .setMobileNumber(options.getMobileNumber())
-                .setMetaData(options.getMetaDataMap())
-                .setToken(options.getCardToken())
-                .build();
-
-        apiService.tokenPreAuth(tokenTransaction)
-                .subscribeOn(scheduler.backgroundThread())
-                .observeOn(scheduler.mainThread())
-                .subscribe(callback(), error());
-    }
-
-}
