+++ judo-sdk/src/test/java/com/judopay/PaymentPresenterTest.java
@@ -3,9 +3,12 @@ package com.judopay;
 import com.google.gson.Gson;
 import com.judopay.arch.Scheduler;
 import com.judopay.model.Card;
+import com.judopay.model.CardToken;
 import com.judopay.model.Currency;
 import com.judopay.model.PaymentRequest;
 import com.judopay.model.Receipt;
+import com.judopay.model.ThreeDSecureInfo;
+import com.judopay.model.TokenRequest;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -21,11 +24,18 @@ import retrofit2.adapter.rxjava.HttpException;
 import rx.Observable;
 
 import static org.mockito.Matchers.any;
+import static org.mockito.Matchers.eq;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 @RunWith(MockitoJUnitRunner.class)
 public class PaymentPresenterTest {
+
+    @Mock
+    CardToken cardToken;
+
+    @Mock
+    ThreeDSecureInfo threeDSecureInfo;
 
     @Mock
     Receipt receipt;
@@ -75,9 +85,7 @@ public class PaymentPresenterTest {
         PaymentPresenter presenter = new PaymentPresenter(transactionCallbacks, apiService, scheduler, gson);
 
         Buffer buffer = new Buffer();
-
         RealResponseBody responseBody = new RealResponseBody(Headers.of("SdkVersion", "5.0"), buffer);
-
         HttpException exception = new HttpException(retrofit2.Response.error(404, responseBody));
 
         when(apiService.payment(any(PaymentRequest.class))).thenReturn(Observable.<Receipt>error(exception));
@@ -112,7 +120,64 @@ public class PaymentPresenterTest {
         verify(transactionCallbacks).onError(any(Receipt.class));
     }
 
-    public Card getCard() {
+    @Test
+    public void shouldPerformTokenPayment() {
+        PaymentPresenter presenter = new PaymentPresenter(transactionCallbacks, apiService, scheduler, gson);
+        when(apiService.tokenPayment(any(TokenRequest.class))).thenReturn(Observable.<Receipt>empty());
+
+        when(cardToken.getToken()).thenReturn("cardToken");
+
+        String consumer = "consumerRef";
+        presenter.performTokenPayment(getCard(), new JudoOptions.Builder()
+                .setCardToken(cardToken)
+                .setConsumerRef(consumer)
+                .setAmount("1.99")
+                .setCurrency(Currency.GBP)
+                .setJudoId("100915867")
+                .build());
+
+        verify(transactionCallbacks).showLoading();
+        verify(apiService).tokenPayment(any(TokenRequest.class));
+    }
+
+    @Test
+    public void shouldShowWebViewWhenAuthWebPageLoaded() {
+        PaymentPresenter presenter = new PaymentPresenter(transactionCallbacks, apiService, scheduler, gson);
+
+        presenter.onAuthorizationWebPageLoaded();
+
+        verify(transactionCallbacks).show3dSecureWebView();
+    }
+
+    @Test
+    public void shouldFinishWhenSuccessfulReceipt() {
+        PaymentPresenter presenter = new PaymentPresenter(transactionCallbacks, apiService, scheduler, gson);
+        String receiptId = "123456";
+
+        when(receipt.isSuccess()).thenReturn(true);
+        when(apiService.complete3dSecure(receiptId, threeDSecureInfo)).thenReturn(Observable.just(receipt));
+
+        presenter.onAuthorizationCompleted(threeDSecureInfo, receiptId);
+
+        verify(transactionCallbacks).onSuccess(eq(receipt));
+        verify(transactionCallbacks).hideLoading();
+    }
+
+    @Test
+    public void shouldShowDeclinedMessageWhenDeclinedReceipt() {
+        PaymentPresenter presenter = new PaymentPresenter(transactionCallbacks, apiService, scheduler, gson);
+        String receiptId = "123456";
+
+        when(receipt.isSuccess()).thenReturn(false);
+        when(apiService.complete3dSecure(receiptId, threeDSecureInfo)).thenReturn(Observable.just(receipt));
+
+        presenter.onAuthorizationCompleted(threeDSecureInfo, "123456");
+
+        verify(transactionCallbacks).onDeclined(eq(receipt));
+        verify(transactionCallbacks).hideLoading();
+    }
+
+    private Card getCard() {
         return new Card.Builder()
                 .setCardNumber("4976000000003436")
                 .setSecurityCode("456")
