+++ judo-sdk/src/test/java/com/judopay/PreAuthPresenterTest.java
@@ -3,9 +3,11 @@ package com.judopay;
 import com.google.gson.Gson;
 import com.judopay.arch.Scheduler;
 import com.judopay.model.Card;
+import com.judopay.model.CardToken;
 import com.judopay.model.Currency;
 import com.judopay.model.PaymentRequest;
 import com.judopay.model.Receipt;
+import com.judopay.model.TokenRequest;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -27,6 +29,9 @@ public class PreAuthPresenterTest {
 
     @Mock
     Receipt receipt;
+
+    @Mock
+    CardToken cardToken;
 
     @Mock
     JudoApiService apiService;
@@ -70,6 +75,25 @@ public class PreAuthPresenterTest {
                 .build());
 
         verify(transactionCallbacks).onError(any(Receipt.class));
+    }
+
+    @Test
+    public void shouldPerformTokenPreAuth() {
+        PreAuthPresenter presenter = new PreAuthPresenter(transactionCallbacks, apiService, new TestScheduler(), new Gson());
+        when(apiService.tokenPreAuth(any(TokenRequest.class))).thenReturn(Observable.<Receipt>empty());
+        when(cardToken.getToken()).thenReturn("cardToken");
+
+        String consumer = "consumerRef";
+        presenter.performTokenPreAuth(getCard(), new JudoOptions.Builder()
+                .setCardToken(cardToken)
+                .setConsumerRef(consumer)
+                .setAmount("1.99")
+                .setCurrency(Currency.GBP)
+                .setJudoId("100915867")
+                .build());
+
+        verify(transactionCallbacks).showLoading();
+        verify(apiService).tokenPreAuth(any(TokenRequest.class));
     }
 
     public Card getCard() {
